[repc] Generate Source and SimpleSource classes
Having the autogenerated Source class provide data members, the
getter/setter implementations and default values was convenient,
but makes it hard to extend/modify the autogenerated type.

With this change, the generated header file contains two classes:
<MyType>Source and <MyType>SimpleSource.  The SimpleSource version
includes the autogenerated implementation code, while the Source
version is pure virtual so it gives you more flexibility but requires
writing more code.

The idea is to make it easy to start with SimpleSource if you are
defining an API, then switch to the pure virtual version when
appropriate.

Change-Id: I794bb79f32e3174ed5bcacf30d07b49129ac2cfb
Reviewed-by: Volker Krause <volker.krause@kdab.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>