Allow .rep API to differ from Source API
This is a Big Change.

Before this change, QtRO allowed either a .rep based implementation
(which defined the API that had to be implemented), or supported using
arbitrary QObjects (which exposed every Property/Signal/Slot in that
class - excluding inherited ones).

This change allows the .rep file to define an arbitrary subset of
Properties/Signals/Slots to be exposed on the QtRO network.  The mapping
is implemented in the same way the compile time Signal/Slot connections
are, which allows compile errors if the APIs are incompatible.  It also
supports the same types of conversions available for connect() calls.

Backwards compatibility is provided by generating a mapping at runtime
that exposes all Property/Signal/Slot features (in the class itself) in
the mapped api.

Change-Id: I8e9f5fc467dddb050355edb6c7b6098a116ce8ce
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>