Enable Source adapter objects
This change just provides the mechanism for using an adapter class for
Source objects.  The change that uses this (to facilitate
QAbstractItemModel Replicas) will be a subsequent patch.  Since the
adapter object is passed to QRemoteObjectSourcePrivate, which is internal,
this patch should pass all existing tests (i.e., no regressions).  Tests
of the adapter functionality will be included with adapter changes.

Change-Id: Ic8ab8f7d00694a0713e17896b5e72efe7d803fa9
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>