Use QMetaType::type() to get the QMetaType ID
QVariant::nameToType() returns the correct type only for Qt built-in
types. For user types, it will always return QVariant::UserType instead
of the actual id of the user type. This causes a segmentation fault when
the user type is constructed from the wrong type-id. QMetaType::type()
instead returns the correct type-id for built-in and user types.

This change also introduces a small test case for slots with user
return types. This is by intention not a Q_PROPERTY to make sure the
method is actually transmitted and invoked on the server side.

Change-Id: I15a391a0adbc4eb6f8b8abd7b04b6f860d0da09c
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>