[Behavior fix] Only expose .rep API
This fixes a behavior issue.  Prior to this patch, creating a Source
QObject from a .rep class and then extending it with Signals, Slots or
Properties to implement the behavior would actually result in forwarding
the derived (i.e., implementation) API, not just the .rep API.

This patch fixes the behavior so only the .rep API is exposed.

Change-Id: Icb8e7d683cd81d402a9db359b4127ce862f2ff31
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>