Always add directory of rep files to INCLUDEPATH
This patch will automatically add the parent directory for each
rep file to INCLUDEPATH. This allows to include headers in rep
files via a relative path and does not require them to be in the same
directory.

Previous attempts by adding PWD to INCLUDEPATH in remoteobjects_repc.prf
did not work, since PWD specifies the full path leading to the directory
containing the current file being parsed. This was the directory of
remoteobjects_repc.prf instead of the pro file calling it in the
first place

Change-Id: Ic82a4661e85cd318aa2ea8c29945fa6f4d831f32
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>