Add reasonable support for QAIM
QAbstractItemModels emit Signals for every role for every cell looked up.
This is not an issue in process, but becomes a burden when every Signal
emitted leads to packets on a socket connection.  This patch creates an
adapter for QAIM Sources, which supports smarter caching and batch queries.

This leverages most of the work done in gerrit CR #101342, but with a
changed public API and using some of the recent QtRO changes, like
PendingCalls, Adapters, and SourceApi selection.

Missing features:
- roleName implementation for QML support.
- column insert, remove, move
- Write support.
- Flags support.
- TreeModel support.

Change-Id: I0eb457c34be192711a8d1bbd66033e98b2ccbfbd
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>