Enhance Registry
This adds support for creating the Registry Host whenever, it no longer
needs to be created before Registry Replicas are used.  It also provides
warnings if a Source is enabled that is already in the Registry.

To support consistency, connections are only made when the Registry is
valid.  This prevents the edge case of a locally hosted source getting
connected to and then having a duplicate in the Registry when it connects.

Change-Id: Iffc21a41a72edf5e3c3b4e80d6de23d0b85c4d49
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>