Increase QRemoteObjectReplica::waitForSource timeout
The previous 100ms are far to short and result here, debug-build
at windows 7, in ~half of the time in timeouts. In case I use the
debugger even always.

So, use a bigger default value of 30s what equals the default
timeout used in e.g. QRemoteObjectPendingCall::waitForFinished and
QLocalSocket::waitForReadyRead.

Change-Id: I24c1d567c4fba2f9c8c7e6254cbfe59ccede13c0
Reviewed-by: Kevin Funk <kevin.funk@kdab.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>