Fix assert for insertRow/removeRow of yet not loaded hierarchies
This also turns the initial Q_ASSERT (only enabled with debug-builds)
into a qFatal (abort also in release-build) since if we hit this
something went very wrong along the stack and we cannot recover
if the caller doesn't handle the case it can go wrong (means uses
the new bool). Note that other code in Qt, like e.g. QDBus, also makes
use of qFatal (to answer the question if this isn't to hard) in case
of fatal errors. An alternate approach could be to turn the bool into
an required argument and make sure all code everywhere does handle
this case.

Change-Id: I7efe5e60d4c30b82f8cfa5f9a3a490fb47336663
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>