Add setName() to Nodes
This is primarily for debugging, as the unit
tests will log messages from many node objects
at once, and this patch includes the node
name in the output.

This doesn't actually give a name to the node.
Instead it sets objectName on the private
Node object as well as the Source control
object.

More objects could use this with their
logging, but this seems like a good start.

Change-Id: Id1aa78dc49a640d9c590eb41f52c7afe5b494ea6
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>