Fix connection between Replica and ReplicaPrivate
This is a set of related bugs.  If you have multiple Replicas of the
same
object, they share a ReplicaPrivate so change packets are only sent once
(to the private), not to each Replica.  This bug manifested itself when
a compiled Replica is acquired first, then a dynamic one.  The
m_numSignals member was not initialized in this case (as the metaobject
was already available from the compiled replica), which caused Signals
to be handled as Slots on the dynamic Replica.

This also makes sure every Replica is connected to the initialized
signal.

Finally, if a Replica connects to the ReplicaPrivate which is already
initialzed, the initialized signal is still emitted.

Change-Id: I384fa89374801163e625c89d28d05a7bb98d06ca
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>