Improve connection handling of local sockets.
If there was a leftover connection file it could be corrupt
or the client could actually connect to the socket, if the
server now tried to connect it would have deleted the connection
and the client wouldn't be able to deal with it. But if the connection
is already valid, we don't need to delete the connection file
but listen to it and the connection will be established as expected.

Change-Id: I5a532d3017ba564413e8e804d99f2e8a8f4a0c78
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>