Fix/Enhance datastream handling
The fix part is that connectionclient and connectionserver
were creating a new QDataStream each time read() was
called.  This could "abandon" bytes if read() finished
with bytes left in the stream.  Now the stream is created
once and kept.

The constant QRemoteObjectPackets::dataStreamVersion was
also created and set to QDataStream::Qt_5_0.  This is
used to consistently set the version of all QDataStreams
used, which will help with compatibility between QtRO
versions.

Change-Id: Idca6d12ec3a5ef07ba6d5a88b00a139c16116b1a
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>