[internal] Don't use QRemoteObjectPacket
This is a large change, but it is the same thing repeated.

QRemoteObjectPacket was a set of types for (de-)serializing
QtRO changes to send over the wire.  But it required
copying the changes to/from the packet before using
those changes.  This goes directly to/from QDataStream
instead.  I think it also makes the InitPacket/InitDynamic
steps a bit easier to read, and should improve the init
performance.

Change-Id: I62b1776858ff74a8532c8f2682aa94ac7248a803
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>