[perf] Pass property index, not name
This is a performance optimization.

Instead of passing the property name over the
wire, we only pass the property index.  Now,
because we need to handle adapters (where the
same property index can from from an adapter
as the Source object, we need a new loopup
function.  Thus the change to repc.

Change-Id: I054dc450a48778274fb475f6487c62637c9f9393
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>