QLALR version of repc
This change moves the repc parser from hand-coded to
a QLALR parser.  It uses a special* lexer/parser header
file that combines regexes** with QLALR to make
simplified grammars.  That is, regular expressions
are the building blocks of the grammar, not more basic
tokens.  It doesn't need lex or a compatible lexer
generator.  It should support full conversion to a
complete (non-regex based) language if necessary, but
incrementally.

The goal is to make repc easier to extend by moving to
a grammar.

*qregexparser.h
**As of Qt 5.5, bootstrap (which is required for repc,
since it is a host tool) includes QRegExp, but not
QRegularExpressions.  So this version of the parser
uses QRegularExpressions in non-bootstrap mode, and
includes a bunch of #defines to work with QRegExp in
bootstrap mode.  THIS IS FRAGILE.  The regex syntax
is different for each library.  Ideally QRE will move
to bootstrap, and these hacks can be removed.

Change-Id: I86f4e764fd22702771d6b00358a0bf2da72658c0
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>