Make connection-factories public
This is a large change, but it is completely
a refactoring of code, not code changes.  The
connectionfactory classes are now public and
describe the connection backends (local and
tcp/ip).

Instead of putting the client code in a set
of files and the server code in a set of files,
we have local_backend and tcpip_backend which
include both client/server code.  The packet
type enums were moved to the global header file
to make them available to additional backends.

Change-Id: I83efc9c1f2013f41f97a4350a31a708f484749e4
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Kevin Funk <kevin.funk@kdab.com>
Reviewed-by: Allen Winter <allen.winter@kdab.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>