Run integration tests for all backends
This changes a fair number of lines of code,
but boils down to two changes:

1) The integrations test is changed to a subdirs
project, so the tests can be repeated for each
backend (and easily extended for new backends).

2) The tests themselves are refactored so that
instead of using class variables for nodes, nodes
are constructed in each test.  This avoids the
potential for errors where sources in nodes are
re-used between tests.

Change-Id: I0fcfa9fb83a4fbaa8331ae8a391fea911738de89
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>