Limit which nodes reconnect works for
There is a reconnect mechanism used when a Node we are
connected to drops off.  This change limits the mechanism
so it only attempts to reconnect to Nodes we have attached
to via connectToNode.  If we are connected to a Registry,
we will see the Node reappear from it, so this keeps each
Node from searching on their own.

Change-Id: I9207066e3f5a7d5d0160f677455c2322b2d110e4
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>