Optionally enable QNX's HAM
QNX's High Availability Manager can detect when processes die.

Currently in QtRO, a Replica's validity is only changed when it
detects the connection has gone down, which for most systems
only happens when it tried to send data.  By using HAM, the QNX
backend can alert the Replica when a connected Host Node dies,
so it can change its validity and possibly attempt to reconnect.

Change-Id: If4062ab70b1d57fa9818ec851bd3b2cbf9dc44cc
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: James McDonnell <jmcdonnell@qnx.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>