Allow Replica property values to persist
Replicas already provide a mechanism to show a default value before the
connection to the Source is established.  This enhancement allows the
.rep PROP definition to include a 'PERSISTED' flag as well, which saves
the property value when the Replica is stopped and restores the value
when it is started again.

Because a shared dll (QtRO) should not know or care where and how values
are persisted, a persistence class needs to be set for the owner Node
for this feature to work.

Change-Id: I5718b024fd268c95d1add2b1be49e6864127782e
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>