[Behavior change] Default to no PROP setter
Changes to QtRO properties by a Replica are inherently asynchronous.
The change needs to be propagated to the Source and then sent back to
all Replica objects.  This is counter-intuitive in many circumstances
(such as QML), where a call to a setter, followed by a call to the
corresponding getter is expected to return the "set" value.

With this change, QtRO will now create an assignProp() SLOT on both the
Source and Replica generated code.  The (virtual) Source method with
default to calling the setter.  The setter will not be available on the
Replica, making Source and Replica QObjects asymmetrical.

Change-Id: Ia1b1d12f59b3b6ebd7dea9ba566f76c2adf904e2
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Kevin Funk <kevin.funk@kdab.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>