Fix SourceApi lookup
The coverity scan pointed to a bug in SourceApi where the intent was for
the static parameter checks (based on the static QObject::connect eval)
to return the list of argument types from the lookup.  It didn't before,
now it does.

Updated repc so autogenerated code includes the same changes.

Coverity-Id: 172053
Coverity-Id: 172054
Coverity-Id: 172060
Change-Id: I8e5ef9df028d5895153cdb2a97636218f64fb618
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>