repc: Remove output file in case of invalid input
This makes repc's behavior more consistent with other generators. If
command-line arguments are correct, input file exists, but the input is
malformed => remove output file. Before this patch, repc generated an
empty file.

On a consecutive make run, this empty file would be considered a valid
output / up-to-date and repc would not be re-run. This patch fixes this,
and makes sure repc is re-run each time, until the input is valid.

Change-Id: I17b20b6f0fb5a7d5b632077ae885540a90049afb
Reviewed-by: Continuous Integration (KDAB) <build@kdab.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>