CMake: Rename files generated by qt5_generate_repc
Align the file names to the names used by the QMake build system. If
you're using a CMake-based build system you need to fix your #includes
of the repc-generated files now.

Changes based on a example:
  test_replica.h -> rep_test_replica.h
  test_source.h -> rep_test_source.h

Note this is a SIC change, but we should do this before the first
tech-preview of QtRemoteObjects.

Change-Id: I8a1a00b5e2b5ee2b125a75a2362bbb0874d8eeab
Task-number: QTBUG-60051
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: Fabrice Salvaire <qt@fabrice-salvaire.fr>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>