Add MODEL keyword to repc
This change enables QAbstractItemModel types to be included in .rep files
as Source/Replica members.  QtRO already handles QAIM types, but with an
overload of enableRemoting() and acquireModel() methods.  If the goal of
the .rep file is to decribe the interface, the model should be part of the
interface as well.

For example, the following will include a QAbstractItemModel in the
forwarded data, passing the display role (Qt::DisplayRole).  Note, it uses
QML roleNames (http://doc.qt.io/qt-5/qabstractitemmodel.html#roleNames)

class Media
{
    ENUM state{Stopped, Paused, Playing}
    PROP(QString currentTrack)
    PROP(state playState)
    MODEL tracks(display)
}

Future TODO - update docs for this feature and include ability to use with
a QItemSelectionModel.  Potentially add something for default values in
the Replica until data is fetched.

Change-Id: I53a365fbc2bdb7fc5d3be19c1fb0d12e743050ee
Reviewed-by: Michael Brasser <michael.brasser@live.com>