QtRemoting: fix test TestModelView::testFlags()
QStandardItemModel::dataChanged() now passes the correct modified roles
to dataChanged(). Since an empty role vector is treated as 'all roles may
have changed', the called slots handled the changed data. Now that the
correct role is passed, the slots filter it out as soon as they don't
care for the given role. Therefore we have to make sure to listen for the
correct roles - in this case Qt::UserRole - 1 which is used by
QStandardItem to store it's flags

Task-number: QTBUG-66061
Change-Id: I47d998f9b253a6555486e62bc9cc68cb92b4cc69
Reviewed-by: Thorbjørn Lund Martsum <tmartsum@gmail.com>