Make (internal) Source objects hierarchical
QRemoteObjectSource is the class used to handle forwarding QObject changes
over QtRO, internally storing the listener pointers and managing the send
and receive of packets when changes occur.  Previously, MODEL and CLASS
objects in a parent class were handled by calling enableRemoting separately
for each nested pointer type, resulting in separate (and independent)
QRemoteObjectSource instances.

This change defines a "Root" object (tied to the object enableRemoting() is
called on), and any nested QObject types are then created as child source
objects of the root object.

The implementation creates an abstract QRemoteObjectSourceBase class, with
QRemoteObjectRootSource and QRemoteObjectSource types derived from it.  Only
the Root objects become "known" types on the bus (and thus only Root objects
can be acquired.

A Private class instance is created by the Root object and the pointer
passed to all child sources, providing a mechanism for sharing needed
information (such as the list of listeners) amonst all members of the
hierarchy.

Change-Id: I632e847cad288b4540c7de6676f791ca2e675f46
Reviewed-by: Konstantin Tokarev <annulen@yandex.ru>
Reviewed-by: Michael Brasser <michael.brasser@live.com>