Allow recursive packetizing of types
We create a new type (QRO_) to allow storing a subclass as a variant.  Thus
when a QObject* property is marshalled, all of the pointed to properties
can be sent as a single variant.  This can be decoded into the proper type
on the replica side.

This required giving Node's private a new QRemoteObjectMetaObjectManager,
a class to handle the building of dynamic metatypes.

Change-Id: Ib83b1edb2fd37dd12defb066d4c2b69509985eaa
Reviewed-by: Michael Brasser <michael.brasser@live.com>