Update registry handling
The change allows the registry to use the automatic reconnect strategy to
repeat the attempts to connect if the registry is not initially available.

It also takes advantage of the registry replica's storage of known types,
to allow replicas and sources to be used before the registry initialization
occurs, thus removing the need for the node's atomic isInitialized
parameter.

Change-Id: I1734988f0c7926ff4fb64dbc9a5fd35fee889e7a
Reviewed-by: Michael Brasser <michael.brasser@live.com>