Support non-CONSTANT pointer types
Prior to supporting nested source objects (internally), there was no way
to support changing pointed to objects.  Thus repc forced MODEL and CLASS
types to be CONSTANT.  This change adds update support, which required a
few internal changes.

First of all, the Q_PROPERTY flags were extended to include a new mode,
SOURCEONLYSETTER.  Having a replica setter for a pointer type makes no
sense (it wouldn't then be part of the API contract, and would fail if
the Node containing the replica dropped off the bus).  So READWRITE didn't
make sense for pointer types.  That left READONLY, but that mode generated
a protected, not public, setter on the Source side (the idea being that
something internal to the source class, not public, would change the
values).  This would limit the functionality of pointer types, as they
could only be updated by inheriting from a Source class.  Adding
SOURCEONLYSETTER addresses this, making the property read-only on the
replica-side, but read/write (with a public setter) on the source-side.
MODEL/CLASS now use SOURCEONLYSETTER.

Another internal impact is that templated acquire calls (where metaobject
data does not need to be sent from the source to define the type) can now
be updated, which requires storing the compile-time generated metaobject
for re-use.  Thus a staticTypes member was added to the node metaoject-
manager.

Several tests (the ones with pointer-to-qobject properties) needed updating
since those properties are not constant anymore, and thus don't have (or
require) constructor parameters for setting the values.

Change-Id: If20cb0fb673d5e91613f9c238b514ed05380f975
Reviewed-by: Michael Brasser <michael.brasser@live.com>