Support externally generated QIODevices
This adds the APIs that will be necessary to create SSL sockets flexibly
(and outside of QtRO) and pass them in.  The integration tests are extended
to show everything works if the tcp/ip connection is created outside QtRO.

The Registry is supported by allowing an "external schema" to be set as the
HostNode's url, which is then used by the registry as the address for any
remoted() source objects.  The client Node calls registerExternalSchema()
with a std::function callback that can create the client-side QIODevice
given the registry provided url.

Change-Id: I0f2d0ea270771e096a787134ef87d537769045f6
Reviewed-by: Michael Brasser <michael.brasser@live.com>