Fix restart/nullptr crash
The nominal case for a QObject pointer is to emit a change when the
value changes, i.e. changing the pointer or setting to nullptr.  No
signal is emitted when the value is set to the same value again.

However, if the server is restarted, it doesn't know what previous
values were sent and will send the current values.  An edge-case, where
the pointer was set to nullptr and nullptr is received again, wasn't
properly handled in QtRO before this fix.

Change-Id: I11443d69476c5b6fecda11e38919542de8a24eb1
Fixes: QTBUG-74221
Reviewed-by: Michael Brasser <michael.brasser@live.com>