Fix a couple of metaObject vs m_metaObject issues
For a replica's private implementation, there are two metaObjects.  The
implementation class itself has a metaObject with the metacall magic shared
amongst all replicas for forwarding calls.  That version is returned by
the private implementations metaObject() call.  There is also the replica
instances m_metaObject, which represents the interface the replica is
copying.  This is retrieved from the m_metaObject data member.

There were a few cases where these were mixed up, so this cleans that up
and tries to document the differences a little better.

Change-Id: I6979e9e997a4e64dc35f91e94108165b50e7715f
Reviewed-by: Michael Brasser <michael.brasser@live.com>