Fix dynamic enums in Props and Signals
These worked with repc generated types, but can fail when going to dynamic
types (such as going through a proxy).

This also makes two changes to help with this change:
It adds a #define to dump what is sent for dynamic type info for
initialization (for debugging the protocol itself) and changes the repc
regexes so that a single character enum name can be used (previously the
parsing would fail if the unless the name was at least two characters).

Change-Id: Ied7219c60a901ef0cb00e06ef44001c6d077af89
Reviewed-by: Michael Brasser <michael.brasser@live.com>