Fix return value propagation through proxies
A slot with return values returns a PendingCall object on the replica.  If
such a replica is used in a proxy, a client would get an invalid QVariant
as the reply.  This was because the proxy would try to return the pending-
call, not wait for the slot return from the original source.  This change
fixes that issue.

Change-Id: I1a25ed49ce51729dde4fa4593845946041493ea1
Reviewed-by: Michael Brasser <michael.brasser@live.com>