More fixes for (dynamic) enums in signals and slots
This addresses two larger issues with enums and several smaller fixes.

First, unlike the recent fix for properties, we need to register enums for
queued connections to work and get signal/slot updates over QtRO to work.
This change adds said registration.

Second, enums are identified with the class they are defined in, so passing
updates through a dynamic object requires converting the typenames.  The
templated enableRemoting methods become interesting here, as the class
name depends on the type passed as a template parameter.  This change also
updates the repc output so enums are properly resolved via the templates.

Smaller fixes include updates to some tests and several places where enums
were converted to/from int.  Note: we need (and have always needed) to
convert enums to ints, since the class name is passed as part of the
variant serialization which cannot be changed without major digging into
qvariant and QDataStream serialization internals.

Change-Id: I0a77f85df6a400a7a44394a05c9c2401bee4e4a8
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>