Eradicate some simple Q_FOREACH loops
All these are relatively easily verified to be safe: They clearly
don't modify the container they iterate over, are already const or
trivially marked as such.

In QRegistrySource::removeServer(), merge with an adjacent loop.

In tst_modelview.cpp, replaced a loop over QHash::keys() + QHash::op[]
with an STL-style loop and it.key()/it.value().

In tst_signature.cpp, replaced a QStringList with a C array of
QLatin1Strings, saving a qAsConst call.

Add some reserve() as a drive-by.

Change-Id: Ibf09371ca8b58437834734c3bab997527219c823
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>