Port from implicit to explicit atomic operations
The old code used the implicit conversions from QAtomicInteger<T> to
T, and vice versa. The semantics of these differ from the ones
std::atomic uses, so we're going to deprecate these, like we did for
load() and store(), too.

This patch fixes some users of these APIs before we deprecate them.

No attempt is made to optimize the atomics use here. This is a
straight port which is functionally identical to the old code.

Change-Id: Ib304a185ed6e34206c9b0eff38cccec844099dd9
Reviewed-by: Lars Knoll <lars.knoll@qt.io>