Fix crash when using QList<*> with a proxy
Qt6 fixes the original test case, as it now generates metatype info at
compile time for container classes of all primitive types.  `subclass.rep`
was modified to include the type QList<MyPOD> as a property, which did
crash in Qt6, because PODs can have their metaObjects created at runtime,
which do not generate container<POD> metatypes.

There are two use-cases which should be supported by a fix.
1) Proxy, where the parameter is stored, but then forward to downstream
   replicas which can be compiled.  In this case, transforming to something
   that works dynamically should support conversion back to the declared
   type.
2) Dynamic replicas being used from QML.  In this case, the dynamic type
   should be made available to QML in a way that is equivalent to the
   declared type.

Dynamic storage and QML support are from the type QtROSequentialContainer.
This class is derived from QVariant list, but stores a few extra members
allow forwarding without losing type information.

Over-the-wire transfer is via a new class (QSQ_), which stores the elements
in a QByteArray based on the valueType, not the container.  Since the type
will always be fully known at the source, we can use metatype detection to
separate out the problematic types (a sequential container of Q_GADGET
types), and convert those to QSQ_ containers for sending.

Note: there is one additional piece needed.  When we send the definition of
a dynamic type (needed for the creation of the dynamic metaobject), we need
to change the type of any member or parameter that is a such a container
to QtROSequentialContainer.

Pick-to: 6.2
Change-Id: Ie2ca53bcbeb4ca0e38c672b5ce951742f625711a
Reviewed-by: Michael Brasser <michael.brasser@live.com>