Fix crash when remoting QML objects with sub-objects
There are two edge cases to watch out for.  First, the property metaType's
metaObject() returns a nullptr, so the property QObject pointer's
metaObject needs to be used instead.  If the QObject is null, the property
is skipped, as there is no way to get the metaobject without introducing
a dependency on QtQuick.

Change-Id: I60ae36dffb30c840858e4d67b9980299c58354b4
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 28093b4ac318209f12189efda9094797226eead1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>