Make file lists in remoteobjects.pro consistent
Fix line continuations of the last list entry, because removing the
empty line between the lists will lead to "interesting" effects.

Remove the $$PWD/ prefix, because that's implict for .pro files.

Remove split into PRIVATE_HEADERS and PUBLIC_HEADERS, because
pro2cmake cannot cope with it, and the distinction is obvious from the
_p.h suffix.

Sort the lists.

Change-Id: I59170f40892193795fdc79167f03b2bfe2fbfe25
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>