Prospective build fix after changes in qtbase's bootstrap library
The binary json conversion code was removed from the bootstrap library,
which broke the repc build. However the part of repc that stopped
compiling was the moc generator code, which repc doesn't use. Therefore
the easiest way to fix the build is to simply remove that code from the
build and delete code that's now unused.

Change-Id: Iee763adc978f98c9bb5bf33817e4cd0cb5d76bc9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>