Fix tests with latest qtbase
The test uses a QVector<int> property, stored in a QVariant and then
compares the variants with QCOMPARE. That relies on the equals
comparators to be registered, which apparently used to be the case due
to some implicit registration. Since that's not the case anymore in
qtbase, let's register them explicitly for the tests to pass.

Change-Id: Ife3abf36f15e761186193b1e734485283db45b2f
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>